% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{check.character}
\alias{check.character}
\title{Check character column}
\usage{
\method{check}{character}(x, n = 20)
}
\arguments{
\item{x}{numeric}

\item{n}{number of random values to show – 20 by default}
}
\description{
Check character column
}
