% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{check.numeric}
\alias{check.numeric}
\title{Check numeric column}
\usage{
\method{check}{numeric}(x)
}
\arguments{
\item{x}{numeric}
}
\value{
No return values - a box/dotplot is produced to examine values
}
\description{
Check numeric column
}
