% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{check_df}
\alias{check_df}
\title{Summarise each column of a data frame}
\usage{
check_df(df)
}
\arguments{
\item{df}{The data frame to be summarised}
}
\value{
Returns \code{NULL}, but outputs figures and summary tables
}
\description{
Summarise each column of a data frame
}
