% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{conditionally}
\alias{conditionally}
\title{Convert a function into a conditional function}
\usage{
conditionally(fun)
}
\arguments{
\item{fun}{Name of function to convert}
}
\value{
Function \code{fun} with an additional logical parameter \code{execute} that toggles whether it is
executed or not.
}
\description{
This function was created by \emph{ablack3} and posted
\href{https://github.com/tidyverse/magrittr/issues/109#issuecomment-371999796}{here} on GitHub.
}
\details{
This function takes another function \code{fun} and adds a logical parameter called \code{execute} that
that specifies whether the function is executed or not.
}
