% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style.R
\name{fairey}
\alias{fairey}
\title{Shepard Fairey colour palette}
\usage{
fairey(num, type = c("discrete", "continuous"))
}
\arguments{
\item{num}{Number of colours required}

\item{type}{\code{discrete} or \code{continuous} colours}
}
\value{
Returns colours palette with \code{num} colours
}
\description{
Shepard Fairey colour palette
}
