% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{strip}
\alias{strip}
\title{Remove all attributes except class.}
\usage{
strip(vec)
}
\arguments{
\item{vec}{A data frame column or vector}
}
\value{
Return \code{vec} with no attributes except \code{class}
}
\description{
Remove all attributes except class.
}
