% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style.R
\name{theme_ocean}
\alias{theme_ocean}
\title{Simple theme for ggplot2 figures}
\usage{
theme_ocean(...)
}
\arguments{
\item{...}{Arguments passed to \code{theme} function of \code{ggplot2}}
}
\description{
Simple theme for ggplot2 figures
}
\seealso{
\code{\link{ggplot2}{theme}}
}
