% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style.R
\name{theme_workhorse}
\alias{theme_workhorse}
\title{Basic theme for ggplot2 figures}
\usage{
theme_workhorse(...)
}
\arguments{
\item{...}{Arguments passed to \code{theme} function of \code{ggplot2}}
}
\description{
Basic theme for ggplot2 figures
}
\seealso{
\code{\link{ggplot2}{theme}}
}
