% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{wm}
\alias{wm}
\title{Calculate proportions of two water masses in a third (target) water mass using a conservative tracer.}
\usage{
wm(t1, t2, tT)
}
\arguments{
\item{t1}{Tracer in first water mass}

\item{t2}{Tracer in second water mass}

\item{tT}{Tracer of target water mass}
}
\value{
Return a data frame containing proportion of two water masses calculated from temperature & salinity.
}
\description{
Calculate proportions of two water masses in a third (target) water mass using a conservative tracer.
}
